+++ package.json
@@ -1,7 +1,21 @@
 {
+  "name": "project-1",
+  "version": "1.0.0",
+  "description": "",
   "type": "module",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "Muhammad",
+  "license": "MIT",
   "dependencies": {
-    "@apollo/client": "^3.7.10",
+    "@apollo/client": "^3.7.11",
+    "@cremafinance/crema-sdk-v2": "^2.4.13",
+    "@orca-so/orca-sdk": "^0.0.4",
+    "@project-serum/serum": "^0.13.65",
+    "@raydium-io/raydium-sdk": "^1.3.0-beta.9",
+    "@solana/spl-token": "^0.3.7",
+    "@solana/web3.js": "^1.75.0",
     "graphql": "^16.6.0",
     "graphql-request": "^5.2.0",
     "kafkajs": "^2.2.4",
